Improve test suite organization
