+++ tests/bootstrap.php
@@ -1,4 +1,3 @@
 <?php
 
 require_once __DIR__.'/../vendor/autoload.php';
-require_once __DIR__.'/Test/IdephixTestCase.php';
