+++ tests/idxrc.php
