+++ tests/unit/Idephix/CommandWrapperTest.php
