+++ tests/unit/Idephix/Config/ConfigTest.php
