+++ tests/unit/Idephix/Config/LazyConfigTest.php
