+++ tests/unit/Idephix/Extension/Deploy/DeployTest.php
@@ -1,7 +1,7 @@
 <?php
 namespace Idephix\Extension\Deploy;
 
-use Idephix\Tests\Test\IdephixTestCase;
+use Idephix\Test\IdephixTestCase;
 use Idephix\Extension\Deploy\Deploy;
 
 class DeployTest extends IdephixTestCase
