+++ tests/unit/Idephix/File/FunctionBasedIdxFileTest.php
