+++ tests/unit/Idephix/SSH/CLISshProxyTest.php
