+++ tests/unit/Idephix/SSH/SshClientTest.php
